+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
 
-- **[UPDATE]** Add `noHorizontalSpacing` prop to disable horizontal padding on `Modal` component.
 [...]
+
+# v37.2.0 (23/07/2020)
+
+- **[UPDATE]** Add `noHorizontalSpacing` prop to disable horizontal padding on `Modal` component.
 - **[NEW]** Added `FilterBar` component
 
 # v37.1.1 (22/07/2020)